chore: move test to proper place
